+++ src/main/java/com/marginallyclever/convenience/SoundSystem.java
@@ -9,9 +9,14 @@ import javax.sound.sampled.Clip;
 
 import com.marginallyclever.convenience.log.Log;
 import com.marginallyclever.makelangelo.makelangeloSettingsPanel.SoundPreferences;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 @Deprecated
 public class SoundSystem {
+
+	private static final Logger logger = LoggerFactory.getLogger(SoundSystem.class);
+
 	static public void playSound(String url) {
 		if (url.isEmpty()) return;
 
@@ -22,7 +27,7 @@ public class SoundSystem {
 			clip.open(inputStream);
 			clip.start();
 		} catch (Exception e) {
-			Log.error(e.getMessage());
+			logger.error("Failed to play sound", e);
 		}
 	}
